@@ -24,7 +24,7 @@ if(POLICY CMP0076)
 endif()
 
 # Set CXX flags
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CXX_EXTRA_FLAGS "-DBOOST_ALL_NO_LIB=1 -DBOOST_NO_CXX14_CONSTEXPR")
 # message(${CMAKE_CXX_COMPILER})
 if (CMAKE_CXX_COMPILER_ID MATCHES "Intel")
